Organize file structure as follow-up for evaluators refactoring
